third_party/WebKit/LayoutTests/http/tests/budget/get-budget-in-service-worker.html

<!DOCTYPE html>
<html>
  <head>
    <title>getBudget succeeds from a Service Worker</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
    <script src="../notifications/resources/test-helpers.js"></script>
    <script src="/js-test-resources/mojo-helpers.js"></script>
    <script src="budget-service-mock.js"></script>
  </head>
  <body>
    <script>
      async_test(function(test) {
        var script = 'resources/instrumentation-service-worker.js';
        var scope = 'resources/scope/' + location.pathname;
        var port;

        budgetServiceMock.then(mock => {
          getActiveServiceWorkerWithMessagePort(test, script, scope)
            .then(function(workerInfo) {
              port = workerInfo.port;
              port.postMessage({command: 'getBudget'});

              port.addEventListener('message', function(event) {
                if (typeof event.data != 'object' || !event.data.command)
                  assert_unreached('Invalid message from the service worker');

                assert_equals(event.data.command, 'getBudget');
                assert_true(event.data.success,
                    'getBudget should succeed. Error message: ' + event.data.errorMessage);
                assert_equals(event.data.budgetAt, TEST_BUDGET_AT);
                assert_equals(event.data.time, TEST_BUDGET_TIME);
                test.done();
              });
            })
          .catch(unreached_rejection(test));
        });
      }, 'getBudget should succeed from Service Worker');
    </script>
  </body>
</html>
